requirements.txt
langchain-openai
streamlit_feedback
tiktoken
langchain-community
streamlit-feedback
beautifulsoup4
